wip: really rough extracting of Base Link to offlineMutation
