+++ packages/offix-offline/src/index.ts
@@ -28,6 +28,7 @@ export * from "./conflicts/strategies/ConflictResolutionStrategy";
 export * from "./conflicts/strategies/ConflictMetaData";
 export * from "./conflicts/state/VersionedState";
 export * from "./conflicts/BaseLink";
+export * from "./conflicts/BaseProcessor";
 export * from "./conflicts/strategies/ConflictResolutionData";
 export * from "./offline/OfflineProcessor";
 export * from "./offline/OfflineError";
